refactor: split Quick Battle apply and round-limit outcome (#4284)
Slice B — move applyQuickBattleResultToGame to quick_battle_apply.dart and
resolveQuickBattleRoundLimitOutcome to quick_battle_resolver_outcome.dart
so quick_battle_resolver.dart owns round-loop orchestration only.
