+++ packages/colonizethis_combat/lib/colonizethis_combat.dart
@@ -13,6 +13,7 @@ export 'src/combat/military_attack_economy.dart';
 export 'src/combat/military_strength.dart';
 export 'src/combat/naval_combat_resolver.dart';
 export 'src/combat/naval_interception.dart';
+export 'src/combat/quick_battle_apply.dart';
 export 'src/combat/quick_battle_input_builder.dart';
 export 'src/combat/quick_battle_resolver.dart';
 export 'src/combat/unopposed_province_capture.dart';
