+++ packages/colonizethis_combat/lib/src/combat/quick_battle_resolver.dart
@@ -1,15 +1,13 @@
 import 'package:colonizethis_combat/src/logging.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import 'package:colonizethis_world/colonizethis_world.dart';
-import 'combat_resolver_support.dart';
 import 'combat_rng.dart';
-import 'combat_survivor_units.dart';
-import 'conflict_detection.dart';
 import 'quick_battle_action_modifiers.dart';
 import 'quick_battle_emplaced_guns.dart';
 import 'quick_battle_resolver_engine.dart';
 import 'quick_battle_resolver_outcome.dart';
+
+export 'quick_battle_apply.dart';
 
 /// Quick Battle resolution pipeline. SPEC/program/quick-battle-resolution.md.
 /// Deterministic for given seed; output feeds same casualty/flip pipeline as auto-resolve.
@@ -202,108 +200,9 @@ QuickBattleResult resolveQuickBattle(
     }
   }
 
-  return _resolveQuickBattleRoundLimitOutcome(
+  return resolveQuickBattleRoundLimitOutcome(
     finisher,
     attGroups: attGroups,
     defGroups: defGroups,
   );
-}
-
-QuickBattleResult _resolveQuickBattleRoundLimitOutcome(
-  QuickBattleFinisher finisher, {
-  required List<QuickBattleGroup> attGroups,
-  required List<QuickBattleGroup> defGroups,
-}) {
-  final input = finisher.input;
-  final finalAttStr =
-      effectiveStrength(attGroups, input.attackerDeployment.laneTerrain) *
-      input.attackerLeaderMultiplier;
-  final finalDefStr =
-      effectiveStrength(defGroups, input.defenderDeployment.laneTerrain) *
-          input.defenderLeaderMultiplier +
-      (finisher.useVirtualEmplaced
-          ? aliveGunStrengthSum(finisher.mutableGuns)
-          : 0.0);
-
-  if (finalAttStr > finalDefStr * 1.2) {
-    return finisher.finish(
-      winner: QuickBattleWinner.attacker,
-      provinceFlips: true,
-    );
-  }
-  if (finalDefStr > finalAttStr * 1.2) {
-    return finisher.finish(
-      winner: QuickBattleWinner.defender,
-      provinceFlips: false,
-    );
-  }
-  return finisher.finish(
-    winner: QuickBattleWinner.mutualExhaustion,
-    provinceFlips: false,
-  );
-}
-
-/// Applies QuickBattleResult to Game: remove casualties, flip province if winner is attacker.
-Game applyQuickBattleResultToGame(
-  Game game,
-  BattleContext ctx,
-  QuickBattleResult result,
-) {
-  final region = game.worldState.regionDataForIdOrThrow(ctx.regionId);
-  final casualtySet = {
-    ...result.attackerCasualties,
-    ...result.defenderCasualties,
-  };
-  final survivingUnits = unitsExcludingCasualtyIds(
-    region.units,
-    casualtySet,
-  ).toList();
-
-  var provinces = region.provinces;
-  if (result.fortDowngradeFromDestroyedEmplaced) {
-    provinces = decrementFortLevelForProvinceIdIfPresent(
-      provinces,
-      ctx.provinceId,
-    );
-  }
-
-  final newRegion = RegionData(provinces: provinces, units: survivingUnits);
-  final newWorldState = game.worldState.updateRegionById(
-    ctx.regionId,
-    (_) => newRegion,
-  );
-
-  var updatedGame = game.withWorldState(newWorldState);
-
-  if (result.provinceFlips &&
-      result.winner == QuickBattleWinner.attacker &&
-      ctx.attackers.isNotEmpty) {
-    final attackerFactionId = ctx.attackers.first.factionId;
-    final row = resolveProvinceRowForOwnershipTransfer(
-      game.worldState,
-      ctx.provinceId,
-    );
-    final oldOwnerId = row?.province.ownerId ?? ctx.defenderFactionId;
-    updatedGame = applyCanonicalSingleProvinceOwnershipTransfer(
-      updatedGame,
-      targetProvinceId: ctx.provinceId,
-      oldOwnerId: oldOwnerId,
-      newOwnerId: attackerFactionId,
-    );
-  }
-
-  updatedGame = updatedGame.copyWith(
-    worldState: reconcileArmiesAfterUnitsChanged(
-      updatedGame.worldState,
-      updatedGame,
-    ),
-  );
-
-  final winnerFactionId = switch (result.winner) {
-    QuickBattleWinner.attacker =>
-      ctx.attackers.isNotEmpty ? ctx.attackers.first.factionId : null,
-    QuickBattleWinner.defender => ctx.defenderFactionId,
-    QuickBattleWinner.mutualExhaustion => null,
-  };
-  return awardWinningGeneralMedal(updatedGame, ctx, winnerFactionId);
 }
